Got stdin/stdout working\! Moved WindowSpawnImpl to main swsh target
