+++ Tests/swshTests/integration-tests/IntegrationTests.swift
@@ -196,7 +196,6 @@ final class IntegrationTests: XCTestCase {
     }
 
     func testWindowsEchoViaGitCompatibility() throws {
-        // XCTAssertEqual(try cmd("C:\\Program Files\\Git\\usr\\bin\\echo.exe", "hello").runString(), "hello")
-        XCTAssertTrue(try cmd("C:\\Program Files\\Git\\usr\\bin\\echo.exe", "hello").runBool())
+        XCTAssertEqual(try cmd("C:\\Program Files\\Git\\usr\\bin\\echo.exe", "hello").runString(), "hello")
     }
 }
